backend/requirements.txt
fastapi
uvicorn
sqlalchemy
PyJWT
bcrypt
slowapi
pytest
jsonschema